renamed sample package
