--- aeron-quick-samples/src/main/java/org/jetc/aeron/quick/samples/annotation_receiver_and_sender/ExamplePojo.java
@@ -1,3 +1,0 @@
-package org.jetc.aeron.quick.samples.annotation_receiver_and_sender;
-
-public record ExamplePojo(String strValue1, long longVal2) {}
