+++ aeron-quick-samples/src/main/java/org/jetc/aeron/quick/samples/basic_send_receive/ExamplePojo.java
@@ -1,0 +1,3 @@
+package org.jetc.aeron.quick.samples.basic_send_receive;
+
+public record ExamplePojo(String strValue1, long longVal2) {}
