+++ aeron-quick-samples/src/main/java/org/jetc/aeron/quick/samples/basic_send_receive/MyGeneralReceiver.java
@@ -1,4 +1,4 @@
-package org.jetc.aeron.quick.samples.annotation_receiver_and_sender;
+package org.jetc.aeron.quick.samples.basic_send_receive;
 
 import org.agrona.concurrent.SleepingMillisIdleStrategy;
 import org.jetc.aeron.quick.AeronQuickFactory;
